Compute distances between grid codes on the plane, not between phases
Before this change, we assumed the readout resolution was a "phase
resolution", a circle in phase space centered on that phase. But with
hexagonal grid cells, a circle in phase space corresponds to an
ellipse in physical space. Said differently, we want to measure
distances on the rhombus, not distances on the [0,1]x[0,1] square.
This change adds that flexibility.