Fix bug that was lurking but not causing problems
This includes a bunch of tests to catch this type of bug, although
these tests don't fail when you revert this change, but they'd catch
the bug if a future change unveils the bug.

The bug is an off-by-one error. It doesn't cause any issues in the
current algorithm, but lots of potential changes would cause this bug
to surface. For example, if you rearrange findGridCodeZeroHelper to
call tryProveGridCodeZeroImpossible before tryFindGridCodeZero, one of
these new tests will fail.